refactor: 프로파일링 API 클라이언트 도메인으로 이전 (#58)
